requirements.txt
# Core packages
flask==2.3.2
flask-cors==3.0.10
flask-socketio==5.3.4
python-engineio==4.3.4
python-socketio==5.8.0
eventlet==0.33.3

# Computer Vision
opencv-python-headless==4.8.0.74
numpy==1.24.3  # Pinned to version compatible with tensorflow-cpu

# Machine Learning
tensorflow-cpu==2.13.0  # Version supporting CAST op version 5
mediapipe==0.10.9  # Latest stable version

# Additional dependencies
protobuf>=3.20.3,<4  # Minimum version for TF while keeping mediapipe compatibility
absl-py==1.4.0
flatbuffers>=2.0  # Required by mediapipe 0.10.9
